refactor: Moved login menu to individual files for better seperation
