+++ src/features/login/LoginButton.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+interface LoginButtonProps {
+	authorizeUrl: string;
+}
+
+const LoginButton: React.FC<LoginButtonProps> = ({ authorizeUrl }) => (
+	<a className="btn btn-primary" href={authorizeUrl}>
+		Login
+	</a>
+);
+
+export default LoginButton;
